+++ Onova.Tests/Extracting/NugetPackageSpecs.cs
@@ -13,16 +13,21 @@ namespace Onova.Tests.Extracting;
 
 public class NugetPackageSpecs : IDisposable
 {
-    private string TempDirPath { get; } = Path.Combine(
-        Directory.GetCurrentDirectory(),
-        $"{nameof(NugetPackageSpecs)}_{Guid.NewGuid()}"
-    );
+    private string TempDirPath { get; } =
+        Path.Combine(
+            Directory.GetCurrentDirectory(),
+            $"{nameof(NugetPackageSpecs)}_{Guid.NewGuid()}"
+        );
 
     public NugetPackageSpecs() => DirectoryEx.Reset(TempDirPath);
 
     public void Dispose() => DirectoryEx.DeleteIfExists(TempDirPath);
 
-    private void CreateNugetPackage(string filePath, string rootDirPath, IReadOnlyDictionary<string, byte[]> entries)
+    private void CreateNugetPackage(
+        string filePath,
+        string rootDirPath,
+        IReadOnlyDictionary<string, byte[]> entries
+    )
     {
         using var zip = ZipFile.Open(filePath, ZipArchiveMode.Create);
 
@@ -36,12 +41,12 @@ public class NugetPackageSpecs : IDisposable
         // Arrange
         var entries = new Dictionary<string, byte[]>
         {
-            ["File1.bin"] = new byte[] {1, 2, 3},
-            ["File2.bin"] = new byte[] {4, 5, 6},
+            ["File1.bin"] = new byte[] { 1, 2, 3 },
+            ["File2.bin"] = new byte[] { 4, 5, 6 },
             ["SubDir1/"] = new byte[0],
-            ["SubDir1/File3.bin"] = new byte[] {7, 8, 9},
+            ["SubDir1/File3.bin"] = new byte[] { 7, 8, 9 },
             ["SubDir1/SubDir2/"] = new byte[0],
-            ["SubDir1/SubDir2/File4.bin"] = new byte[] {10, 11, 12}
+            ["SubDir1/SubDir2/File4.bin"] = new byte[] { 10, 11, 12 }
         };
 
         var packageFilePath = Path.Combine(TempDirPath, "Package.nupkg");