+++ Onova.Updater/Utils/DirectoryEx.cs
@@ -1,58 +1,63 @@
 ﻿using System;
-using System.ComponentModel;
+using System.Collections.Generic;
 using System.IO;
-using System.Runtime.InteropServices;
-using Onova.Updater.Native;
+using System.Linq;
+using Onova.Updater.Utils.Extensions;
 
 namespace Onova.Updater.Utils;
 
 internal static class DirectoryEx
 {
-    public static IDisposable Lock(
-        string dirPath,
-        FileAccess access = FileAccess.ReadWrite,
-        FileShare share = FileShare.None)
-    {
-        if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            return Disposable.Null;
-
-        var handle = NativeMethods.CreateFile(
-            dirPath,
-            access,
-            share,
-            0,
-            FileMode.Open,
-            (FileAttributes)0x02000000,
-            0
-        );
-
-        if (handle.IsInvalid)
-        {
-            handle.Dispose();
-            throw new Win32Exception();
-        }
-
-        return handle;
-    }
-
+    // Performs a recursive copy of the source directory to the destination directory as an atomic operation
     public static void Copy(string sourceDirPath, string destDirPath, bool overwrite = true)
     {
-        Directory.CreateDirectory(destDirPath);
+        // Acquire locks in the destination directory for each file from the source directory
+        var sourceStreams = new List<FileStream>();
+        var destStreams = new List<FileStream>();
 
-        // Copy files
-        foreach (var sourceFilePath in Directory.GetFiles(sourceDirPath))
+        try
         {
-            var destFileName = Path.GetFileName(sourceFilePath);
-            var destFilePath = Path.Combine(destDirPath, destFileName);
-            File.Copy(sourceFilePath, destFilePath, overwrite);
+            foreach (
+                var sourceFilePath in Directory.GetFiles(
+                    sourceDirPath,
+                    "*",
+                    SearchOption.AllDirectories
+                )
+            )
+            {
+                sourceStreams.Add(
+                    File.Open(sourceFilePath, FileMode.Open, FileAccess.Read, FileShare.Read)
+                );
+
+                var destFilePath = Path.Combine(
+                    destDirPath,
+                    PathEx.GetRelativePath(sourceDirPath, sourceFilePath)
+                );
+
+                Directory.CreateDirectory(Path.GetDirectoryName(destFilePath) ?? destDirPath);
+
+                destStreams.Add(
+                    File.Open(
+                        destFilePath,
+                        overwrite ? FileMode.OpenOrCreate : FileMode.CreateNew,
+                        FileAccess.ReadWrite,
+                        FileShare.None
+                    )
+                );
+            }
+
+            // Copy data from the source files to the destination files
+            foreach (var (sourceStream, destStream) in sourceStreams.Zip(destStreams))
+            {
+                sourceStream.CopyTo(destStream);
+
+                // Truncate the destination file if the source file is shorter
+                destStream.SetLength(sourceStream.Length);
+            }
         }
-
-        // Copy subdirectories recursively
-        foreach (var sourceSubDirPath in Directory.GetDirectories(sourceDirPath))
+        finally
         {
-            var destSubDirName = Path.GetFileName(sourceSubDirPath);
-            var destSubDirPath = Path.Combine(destDirPath, destSubDirName);
-            Copy(sourceSubDirPath, destSubDirPath, overwrite);
+            sourceStreams.Concat(destStreams).Cast<IDisposable>().DisposeAll();
         }
     }
 }