+++ Onova/Exceptions/LockFileNotAcquiredException.cs
@@ -11,7 +11,7 @@ public class LockFileNotAcquiredException : Exception
     /// Initializes an instance of <see cref="LockFileNotAcquiredException" />.
     /// </summary>
     public LockFileNotAcquiredException()
-        : base("Could not acquire a lock file. Most likely, another instance of this application currently owns the lock file.")
-    {
-    }
+        : base(
+            "Could not acquire a lock file. Most likely, another instance of this application currently owns the lock file."
+        ) { }
 }